+++ apps/common/src/main/resources/db/migration/canton-network/postgres/stable/V053__validator_liveness_weight.sql
@@ -1,0 +1,5 @@
+-- Copyright (c) 2024 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+-- SPDX-License-Identifier: Apache-2.0
+
+-- For ValidatorLivenessActivityRecord weight field
+ALTER TABLE dso_acs_store ADD COLUMN validator_liveness_weight numeric;