+++ docs/src/release_notes.rst
@@ -5,6 +5,19 @@
 
 .. _release_notes:
 
+Integration branch
+------------------
+
+- Daml
+
+  - `ValidatorLicense` now has `weight`, and `kind` fields
+
+- SV
+
+  - `dso_acs_store` requires migration to add a column for storing `ValidatorLivenessActivityRecord` weight
+
+  - SVs can grant `ValidatorLicense` to any arbitrary party. See `CIP <https://github.com/global-synchronizer-foundation/cips/blob/main/cip-0073/cip-0073.md>` for the motivation.
+
 .. release-notes:: 0.5.3
 
   Note: 0.5.2 mistakingly introduced default pruning for Canton participants and should be skipped in favor of 0.5.3.